requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.4
django-allauth==0.61.1
django-bootstrap4==24.1
django-crispy-forms==2.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
gunicorn==21.2.0
idna==3.7
oauthlib==3.2.2
packaging==24.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
